Cleaned Up! And Added Anims with TickTo
There is a matching commit in the AIR pixie packer.  This commit cleans
up the original pixiecomponent class, and the associated animations.
The usable component is PixieAnimator.java and requires
PixieHeader.java.
The demo animation now makes use of reading its TickTo value from the
pixie header. Although this value is currently not customizable in the
AIR build, it will be in the future.

This change allows the speed of the animation to be 'interpretted' at
runtime, meaning an animator working on say a progress spinning
animation, could customize the speed on the sheet itself without ever
needing to touch code or a sidecar XML file!!