Relevant Demo
Turned the 10 frames blue/yellow bars into pixies they now appear in the
demo when you run it.

Turns out my lazy alphabet sorting does the oldschool number thing where
the order is:
1, 11, 2, 3, 4, 449, 5, 6
We need:
1, 2, 3, 4, 5, 6, 11, 449

This will need to be changed for file names to work as expected, as a
work around just use letters, a, b, c, where aa falls between a and b.