Added 50 Frame MultiCol Usage Sample
Tried it out with 94 frames for each, all packed into one sheet (~400
frames in one sheet) at the size we have is impossible (~60-80MB in
RAM).

Normally this is not an issue.  On 2.2 this is an issue. We run 2.2
(froyo)
We can make due, by reducing the frame count, or reducing the size. As
you can see in this commit we can comfortably use ~26MB of memory and
still maintain suuuuuuuper buttery animations across 4 colors